Move code to AwsAuroraPgvectorExtensionCreatorNestedStack
